make the plan wait properly for interactive start
moved the helper to a place from where it can be used in
the executable as well as in the testcases.
